+++ Compiler/Tests/Choose.cs
@@ -14,5 +14,17 @@ namespace Exodrifter.Rumor.Compiler.Tests
 			var node = Compiler.Choose(state);
 			Assert.AreEqual(new ChooseNode(), node);
 		}
+
+		[Test]
+		public static void ChooseTimeoutSuccess()
+		{
+			var state = new ParserState("choose in 2 seconds or jump foobar", 4);
+
+			var node = Compiler.Choose(state);
+			Assert.AreEqual(
+				new ChooseNode(new NumberLiteral(2), "foobar"),
+				node
+			);
+		}
 	}
 }